refactor: split lesson editor state into smaller hooks
